+++ lib/app/models/user_settings.dart
@@ -58,7 +58,6 @@ class DefaultUserSettings {
   static const interruptionSkipBackward = Duration.zero;
   static const interruptionLongSkipThreshold = Duration(seconds: 30);
   static const interruptionLongSkipBackward = Duration.zero;
-  static const autoplay = true;
 }
 
 @freezed
@@ -184,8 +183,6 @@ sealed class UserSettings with _$UserSettings {
 
     @Default(DefaultUserSettings.interruptionLongSkipBackward)
     Duration interruptionLongSkipBackward,
-
-    @Default(DefaultUserSettings.autoplay) bool autoplay,
   }) = _UserSettings;
 
   factory fromJson(Map<String, dynamic> json) => _$UserSettingsFromJson(json);